Fix NameError: name 'this' is not defined
